[Refactor] #9 UIViewController extension 추가 및 적용
